requirements.txt
flask
flask-cors
opencv-python-headless
numpy
tensorflow-cpu==2.15.0
gunicorn